% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eNoseMeasurement.R
\name{eNoseMeasurement}
\alias{eNoseMeasurement}
\title{R6 Class representing a measurement}
\description{
R6 Class representing a measurement
}
\details{
For details ...
}
\examples{

rnorm(100)

}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{meta.data}}{meta data}

\item{\code{data_wide}}{data in wide format}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-eNoseMeasurement-clone}{\code{eNoseMeasurement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eNoseMeasurement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-eNoseMeasurement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{eNoseMeasurement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
