\name{odoR-package}
\alias{odoR-package}
\docType{package}
\title{
  odoR ...
}
\description{
  This package provides functions to analyse ...
}
\details{
  \tabular{ll}{
    Package: \tab odoR\cr
    Type: \tab Package\cr
    License: \tab GPL (>= 2)\cr
  }

The package odoR helps with the following ...

}
\author{
Kornelius Rohmeyer

Maintainer: Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\references{
Technical Manual Smell Inspector, SmartNanotubes
\url{https://smart-nanotubes.com/wp-content/uploads/2023/07/SNT_Manual_Smell-Inspector_V1.3.1.pdf}
}
\keyword{ package }
\keyword{ hplot }
\keyword{ IO }
\examples{
rnorm(100)
\dontrun{
rnorm(100)}
rnorm(100)
}
