% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odorDataset.R
\name{odorDataset}
\alias{odorDataset}
\title{R6 Class representing a data set of measurements}
\description{
R6 Class representing a data set of measurements

R6 Class representing a data set of measurements
}
\details{
R6 Class representing a data set of measurements

For details ...
}
\examples{

rnorm(100)

}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{measurements}}{list of measurements}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-odorDataset-new}{\code{odorDataset$new()}}
\item \href{#method-odorDataset-add_measurement}{\code{odorDataset$add_measurement()}}
\item \href{#method-odorDataset-spider_plot}{\code{odorDataset$spider_plot()}}
\item \href{#method-odorDataset-get_features}{\code{odorDataset$get_features()}}
\item \href{#method-odorDataset-pca}{\code{odorDataset$pca()}}
\item \href{#method-odorDataset-clone}{\code{odorDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-odorDataset-new}{}}}
\subsection{Method \code{new()}}{
Create a odorDataset object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorDataset$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `eNoseMeasurement` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorDataset-add_measurement"></a>}}
\if{latex}{\out{\hypertarget{method-odorDataset-add_measurement}{}}}
\subsection{Method \code{add_measurement()}}{
Model the baseline (default: as a polynomial of degree 2)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorDataset$add_measurement(measurement, label = NA, concentration = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measurement}}{Measurement of type odorMeasurement to add.}

\item{\code{label}}{Optional label of sample that was measured.}

\item{\code{concentration}}{Optional numerical value of sample concentration that was measured.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorDataset-spider_plot"></a>}}
\if{latex}{\out{\hypertarget{method-odorDataset-spider_plot}{}}}
\subsection{Method \code{spider_plot()}}{
Draws combined spiderplot of all features in this data set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorDataset$spider_plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorDataset-get_features"></a>}}
\if{latex}{\out{\hypertarget{method-odorDataset-get_features}{}}}
\subsection{Method \code{get_features()}}{
Getter method for the features. Calculates the features if they are not yet existing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorDataset$get_features()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorDataset-pca"></a>}}
\if{latex}{\out{\hypertarget{method-odorDataset-pca}{}}}
\subsection{Method \code{pca()}}{
Principle component analysis of the features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorDataset$pca()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-odorDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
