% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odorMeasurement.R
\name{odorMeasurement}
\alias{odorMeasurement}
\title{R6 Class representing a measurement}
\description{
R6 Class representing a measurement

R6 Class representing a measurement
}
\details{
R6 Class representing a measurement

For details ...
}
\examples{

rnorm(100)

}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{meta.data}}{meta data}

\item{\code{data_wide}}{data in wide format}

\item{\code{data_long}}{data in long format}

\item{\code{baseline_data}}{baseline data as given in the header. Perhaps not that useful, see the discussion about falling response values.}

\item{\code{change}}{...}

\item{\code{name}}{...}

\item{\code{start}}{...}

\item{\code{stop}}{...}

\item{\code{start.probe}}{...}

\item{\code{stop.probe}}{...}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{meta.data}}{meta data}

\item{\code{start.probe}}{...}

\item{\code{stop.probe}}{...}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-odorMeasurement-new}{\code{odorMeasurement$new()}}
\item \href{#method-odorMeasurement-clone}{\code{odorMeasurement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorMeasurement-new"></a>}}
\if{latex}{\out{\hypertarget{method-odorMeasurement-new}{}}}
\subsection{Method \code{new()}}{
Create a eNoseMeasurement object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorMeasurement$new(
  meta_data = NULL,
  data_wide = NULL,
  data_long = NULL,
  baseline_data = NULL,
  change = NULL,
  name = NULL,
  start = NULL,
  stop = NULL,
  start_probe = NULL,
  stop_probe = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meta_data}}{meta data}

\item{\code{data_wide}}{data in wide format}

\item{\code{data_long}}{data in long format}

\item{\code{baseline_data}}{baseline data as given in the header. Perhaps not that useful, see the discussion about falling response values.}

\item{\code{change}}{...}

\item{\code{name}}{...}

\item{\code{start}}{...}

\item{\code{stop}}{...}

\item{\code{start_probe}}{...}

\item{\code{stop_probe}}{...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `eNoseMeasurement` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorMeasurement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-odorMeasurement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorMeasurement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
