% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.odor}
\alias{plot.odor}
\title{Plots the data of a single data set.}
\usage{
\method{plot}{odor}(x, show_baseline = FALSE)
}
\arguments{
\item{x}{object to plot}

\item{show_baseline}{logical whether to show the baseline or not}
}
\value{
A ggplot2 graph object.
}
\description{
Plots the data of a single data set.
}
\details{
Plots the data over the time. For each sensor type (base and features) as well as for temperature and humidity there will be a facet showing the corresponding data.
The events of the measurement will be shown as vertical lines.
}
\examples{

rnorm(100)

}
\seealso{
\code{\link{ggplot}}
}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\keyword{hplot}
