% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.eNase}
\alias{read.eNase}
\title{Reads the data of a single data set.}
\usage{
read.eNase(filename)
}
\arguments{
\item{filename}{The name of the file which the data are to be read from.}
}
\value{
A ggplot2 graph object.
}
\description{
Reads the data of a single data set.
}
\details{
For details ...
}
\examples{

rnorm(100)

}
\seealso{
\code{\link{ggplot}}
}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\keyword{IO}
