% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.odor}
\alias{read.odor}
\title{Reads the data of a single data set.}
\usage{
read.odor(filename, warn = TRUE)
}
\arguments{
\item{filename}{The name of the file which the data are to be read from.}

\item{warn}{Boolean whether warnings should be given about missing events and older firmware.}
}
\value{
An \code{\link{odorDataset}} S6 object.
}
\description{
Reads the data of a single data set. Currently only csv files from Smell Inspector from Smart Nanotubes are supported.
}
\details{
Warnings are given when no events are recorded and when the firmware is rather old.
The later could result in channels having constant readings due to sensor values out of the supported range.
}
\examples{

rnorm(100)

}
\seealso{
\code{\link{odorDataset}}
}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\keyword{IO}
