% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odorMeasurement.R
\name{odorMeasurement}
\alias{odorMeasurement}
\title{R6 Class representing a measurement}
\description{
R6 Class representing a measurement

R6 Class representing a measurement
}
\details{
R6 Class representing a measurement

For details ...
}
\examples{

rnorm(100)

}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{meta_data}}{meta data}

\item{\code{data_wide}}{data in wide format}

\item{\code{data_long}}{data in long format}

\item{\code{baseline_data}}{baseline data as given in the header. Perhaps not that useful, see the discussion about falling response values.}

\item{\code{response}}{data.frame of the calculated responses from this measurement for each sensor}

\item{\code{name}}{Name of the measurement. If read from a csv file the value that was in the included JSON meta data in the field name.}

\item{\code{start}}{POSIXct object representing the start time of the measurement}

\item{\code{stop}}{POSIXct object representing the stop/end time of the measurement}

\item{\code{start_exposure}}{POSIXct object representing the start time of exposure to a test substance}

\item{\code{stop_exposure}}{POSIXct object representing the end time of exposure to a test substance}

\item{\code{base_line_models}}{list of models to fit the base line}

\item{\code{corrected_data_long}}{baseline corrected or residualized data}

\item{\code{features}}{Summarized responses by features}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-odorMeasurement-new}{\code{odorMeasurement$new()}}
\item \href{#method-odorMeasurement-model_baseline}{\code{odorMeasurement$model_baseline()}}
\item \href{#method-odorMeasurement-residualize}{\code{odorMeasurement$residualize()}}
\item \href{#method-odorMeasurement-correct_for_baseline}{\code{odorMeasurement$correct_for_baseline()}}
\item \href{#method-odorMeasurement-calculate_response}{\code{odorMeasurement$calculate_response()}}
\item \href{#method-odorMeasurement-calculate_features}{\code{odorMeasurement$calculate_features()}}
\item \href{#method-odorMeasurement-get_features}{\code{odorMeasurement$get_features()}}
\item \href{#method-odorMeasurement-spider_plot}{\code{odorMeasurement$spider_plot()}}
\item \href{#method-odorMeasurement-clone}{\code{odorMeasurement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorMeasurement-new"></a>}}
\if{latex}{\out{\hypertarget{method-odorMeasurement-new}{}}}
\subsection{Method \code{new()}}{
Create a odorMeasurement object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorMeasurement$new(
  meta_data = NULL,
  data_wide = NULL,
  data_long = NULL,
  baseline_data = NULL,
  name = NULL,
  start = NULL,
  stop = NULL,
  start_exposure = NULL,
  stop_exposure = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meta_data}}{meta data}

\item{\code{data_wide}}{data in wide format}

\item{\code{data_long}}{data in long format}

\item{\code{baseline_data}}{baseline data as given in the header. Perhaps not that useful, see the discussion about falling response values.}

\item{\code{name}}{Name of the measurement as character string}

\item{\code{start}}{POSIXct object representing the start time of the measurement}

\item{\code{stop}}{POSIXct object representing the stop/end time of the measurement}

\item{\code{start_exposure}}{POSIXct object representing the start time of exposure to a test substance}

\item{\code{stop_exposure}}{POSIXct object representing the end time of exposure to a test substance}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `odorMeasurement` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorMeasurement-model_baseline"></a>}}
\if{latex}{\out{\hypertarget{method-odorMeasurement-model_baseline}{}}}
\subsection{Method \code{model_baseline()}}{
Model the baseline (default: as a polynomial of degree 2)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorMeasurement$model_baseline(
  use_data_after_measurement_for_baseline = TRUE,
  method = "polynomial",
  polynomial_degree = 2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{use_data_after_measurement_for_baseline}}{Should the data after the exposure to the odor also be used?
Default is TRUE to reduce extrapolation errors. If there are still major effects from the odor after the measurement, it could make sense to set this to FALSE.}

\item{\code{method}}{Either exponential, polynomial or power-law.}

\item{\code{polynomial_degree}}{Only used if method="polynomial". Degree of polynomial to fit. Default is 2.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorMeasurement-residualize"></a>}}
\if{latex}{\out{\hypertarget{method-odorMeasurement-residualize}{}}}
\subsection{Method \code{residualize()}}{
Residualize.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorMeasurement$residualize(include_factors = c("humidity", "temperature"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_factors}}{factors to include as explaining variables. Default is humidity and temperature.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorMeasurement-correct_for_baseline"></a>}}
\if{latex}{\out{\hypertarget{method-odorMeasurement-correct_for_baseline}{}}}
\subsection{Method \code{correct_for_baseline()}}{
Correct the data by subtracting the baseline.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorMeasurement$correct_for_baseline()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorMeasurement-calculate_response"></a>}}
\if{latex}{\out{\hypertarget{method-odorMeasurement-calculate_response}{}}}
\subsection{Method \code{calculate_response()}}{
Calculate the responses for each sensor as the mean of the values of the corrected data (if existing - otherwise the raw data)
between the start_exposure and stop_exposure events.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorMeasurement$calculate_response()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorMeasurement-calculate_features"></a>}}
\if{latex}{\out{\hypertarget{method-odorMeasurement-calculate_features}{}}}
\subsection{Method \code{calculate_features()}}{
Calculate features by summarizing the responses by feature.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorMeasurement$calculate_features()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorMeasurement-get_features"></a>}}
\if{latex}{\out{\hypertarget{method-odorMeasurement-get_features}{}}}
\subsection{Method \code{get_features()}}{
Getter method for the features. Calculates the features if they are not yet existing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorMeasurement$get_features()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorMeasurement-spider_plot"></a>}}
\if{latex}{\out{\hypertarget{method-odorMeasurement-spider_plot}{}}}
\subsection{Method \code{spider_plot()}}{
Draws a radar chart of the features.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorMeasurement$spider_plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-odorMeasurement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-odorMeasurement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{odorMeasurement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
