% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eNase}
\alias{plot.eNase}
\title{Plots the data of a single data set.}
\usage{
\method{plot}{eNase}(x)
}
\value{
A ggplot2 graph object \code{\link{graphMCP}} or \code{\link{entangledMCP}}.
}
\description{
Plots the data of a single data set.
}
\details{
For details ...
}
\examples{

rnorm(100)

}
\seealso{
\code{\link{ggplot}}
}
\author{
Kornelius Rohmeyer \email{rohmeyer@small-projects.de}
}
\keyword{hplot}
